requirements.txt
streamlit>=1.24.0
openai>=1.0.0
python-dotenv>=0.19.0
difflib